Utils: move package utils/heart to internal/heart
